+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.9.31008](https://github.com/springraaar/skylobby/releases/tag/0.9.31008) - 2025-10-XX
 
 - Reload default preset styles on startup as they may have been updated
+- Avoid redundant tasks in some situations
+- Fixed rapid download for "MechCommander : Legacy"
 
 ## [0.9.31007](https://github.com/springraaar/skylobby/releases/tag/0.9.31007) - 2025-10-08